# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Support Ticket SLA Processing System.",
        "title": "Support Ticket SLA Processing System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login with username and password through Keycloak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Invalid username or password",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/ticket-events/import": {
            "post": {
                "description": "Import ticket events in batch using worker pool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-events"
                ],
                "summary": "Import ticket events",
                "parameters": [
                    {
                        "description": "Import ticket events request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tickets": {
            "get": {
                "description": "Get a list of support tickets with optional filters by status, priority and assignee ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "List tickets",
                "parameters": [
                    {
                        "enum": [
                            "new",
                            "assigned",
                            "in_progress",
                            "resolved",
                            "closed",
                            "cancelled"
                        ],
                        "type": "string",
                        "description": "Filter by ticket status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "type": "string",
                        "description": "Filter by ticket priority",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by assignee ID",
                        "name": "assignee_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List tickets successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new support ticket with title, description, requestor, assignee and priority.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a new ticket",
                "parameters": [
                    {
                        "description": "Create ticket request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTicketReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body or invalid priority",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Get support ticket detail by ticket ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get ticket detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get ticket detail successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid ticket ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tickets/{id}/status": {
            "patch": {
                "description": "Update support ticket status by ticket ID. The status transition must follow the required ticket status flow.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Update ticket status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ticket status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket status updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body, invalid ticket ID or invalid status transition",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Priority": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high"
            ],
            "x-enum-varnames": [
                "PriorityLow",
                "PriorityMedium",
                "PriorityHigh"
            ]
        },
        "domain.TicketStatus": {
            "type": "string",
            "enum": [
                "new",
                "assigned",
                "in_progress",
                "resolved",
                "closed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusNew",
                "StatusAssigned",
                "StatusInProgress",
                "StatusResolved",
                "StatusClosed",
                "StatusCancelled"
            ]
        },
        "dto.CreateTicketReq": {
            "type": "object",
            "required": [
                "priority",
                "requestor_id",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 5000
                },
                "priority": {
                    "$ref": "#/definitions/domain.Priority"
                },
                "requestor_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateStatusReq": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.TicketStatus"
                }
            }
        }
    }
}